refactor: refactor code-embedded functionality to a component
